feat: 添加pnpm、vitest依赖，添加ts配置
