+++ README.md
@@ -1,5 +1,45 @@
-# Computer Science Notes
+---
+home: true
+title: 笔记啦
 
-![GitHub Workflow Status](https://img.shields.io/github/workflow/status/with-rainbow/notes/docs)
+heroText: Notes
+tagline: 为了方便自己管理学习笔记，建立了这个网站，希望对大家也能有帮助。
+actions:
+  - text: 阅读笔记
+    link: /#概述
+    type: primary
+  - text: 个人简介
+    link: /about/
+    type: secondary
+features:
+  - title: 不要追求完美
+    details: 能使用稳定版的，不要使用beta版，费时间。
+  - title: 不要追求完美
+    details: 老想着造轮子是不行的。
+  - title: 不要追求完美
+    details: 不要急于求成。
+footer: Apache 2.0 Licensed | Copyright © 2022-present Rainbow Watcher
+---
 
-this repository is under construction.
+> 笔记是对博客、以及阅读过的书中内容的整理和归纳。
+
+<!-- more -->
+
+# 概述
+
+在计算机领域，每个人都需要花相当多的时间来学习该领域的基础知识，而且在学习的过程中经常会遇到之前学过的内容，一段时间后就会忘记，再次拾起来的过程是比较消耗时间的，为了避免这种情况出现，在这里把各个领域的知识记录下来，归纳成一个完整的知识网络。
+
+## 推荐
+
+### 在线工具
+
+- [菜鸟在线工具集](https://c.runoob.com/)
+- [Tabnine 代码案例搜索](https://www.tabnine.com/code)
+
+## 大纲
+
+笔记包含哪些内容
+
+## 未来
+
+将要添加的内容
