+++ package.json
@@ -8,16 +8,22 @@
   "license": "Apache-2.0",
   "private": false,
   "scripts": {
+    "build": "tsc --build",
     "setup": "ts-node scripts/setup.ts",
-    "docs:dev": "vuepress dev docs --clean-cache",
-    "docs:build": "vuepress build docs --clean-cache"
+    "docs:dev": "vuepress dev --clean-cache",
+    "docs:build": "vuepress build --clean-cache",
+    "test": "vitest",
+    "coverage": "vitest run --coverage"
   },
   "devDependencies": {
     "@vuepress/plugin-search": "^2.0.0-beta.36",
+    "husky": "^7.0.4",
     "markdown-it": "^12.3.2",
     "prettier": "2.6.2",
     "ts-node": "^10.7.0",
+    "tsconfig-vuepress": "^0.1.0",
     "typescript": "^4.6.2",
+    "vitest": "^0.10.2",
     "vuepress": "^2.0.0-beta.38",
     "vuepress-plugin-copy-code2": "^2.0.0-beta.25",
     "vuepress-plugin-md-enhance": "^2.0.0-beta.24"
