+++ packages/plugin-auto-navbar/src/node/autoNavbarPlugin.ts
@@ -1,0 +1,24 @@
+import type { Plugin } from "@vuepress/core"
+// import { path } from "@vuepress/utils"
+
+// export interface Hooks {
+//   onInitialized: LifeCycleHook
+//   onPrepared: LifeCycleHook
+//   onWatched: LifeCycleHook<[watchers: Closable[], restart: () => Promise<void>]>
+//   onGenerated: LifeCycleHook
+//   extendsMarkdownOptions: ExtendsHook<MarkdownOptions>
+//   extendsMarkdown: ExtendsHook<Markdown>
+//   extendsPageOptions: ExtendsHook<PageOptions>
+//   extendsPage: ExtendsHook<Page>
+//   extendsBundlerOptions: ExtendsHook<any>
+//   clientAppEnhanceFiles: ClientFilesHook
+//   clientAppRootComponentFiles: ClientFilesHook
+//   clientAppSetupFiles: ClientFilesHook
+//   alias: ReturnObjectHook
+//   define: ReturnObjectHook
+// }
+
+export const autoNavbarPlugin = (): Plugin => ({
+  name: "@with-rainbow/plugin-auto-navbar",
+  // onInitialized:
+})
