+++ packages/plugin-auto-navbar/src/node/index.ts
@@ -1,0 +1,3 @@
+import { autoNavbarPlugin } from "./autoNavbarPlugin"
+
+export default autoNavbarPlugin
