+++ packages/theme-bys/tsconfig.esm.json
@@ -1,0 +1,15 @@
+{
+  "extends": "../../../tsconfig.base.json",
+  "compilerOptions": {
+    "module": "ES2020",
+    "rootDir": "./src",
+    "outDir": "./lib",
+    "types": [
+      "@vuepress/client/types"
+    ]
+  },
+  "include": [
+    "./src/client",
+    "./src/shared"
+  ]
+}
