+++ scripts/setup.ts
@@ -5,9 +5,9 @@ import markdownIt from "markdown-it"
 import {
   SidebarConfigObject,
   SidebarConfigArray,
-  NavLink,
 } from "@vuepress/theme-default"
-import Token from "markdown-it/lib/token"
+import { NavbarConfig } from "vuepress"
+import { getMDFileTitles } from "../packages/plugin-auto-navbar/src/node/mdUtil"
 
 const log = (...msg: any) => {
   console.log(...msg)
@@ -22,14 +22,16 @@ const DEFAULT_IGNORED_FILES = [".DS_Store", ...HOME_PAGE_NAME]
  * 获取导航栏配置对象
  */
 function getNavBarConfig(folders: string[]) {
-  let navbar_object: NavLink[] = folders.map((value) => {
+  let navbar_object: NavbarConfig = folders.map(value => {
     const filePath = path.resolve(DOCS_PATH, value)
     const homePageFileName = getHomePageFileName(filePath)
     const homePageFilePath = path.resolve(filePath, homePageFileName)
     const title = getMDFileTitle(homePageFilePath)
+    log(getConfigArray(value))
     return {
       text: title ?? value,
       link: `/${value}/`,
+      children: getConfigArray(value),
     }
   })
   return navbar_object
@@ -41,7 +43,7 @@ function getNavBarConfig(folders: string[]) {
 function initSidebarConfigObject(folders: string[]): SidebarConfigObject {
   let configObject: { [k: string]: SidebarConfigArray } = {}
 
-  folders.map((val) => {
+  folders.map(val => {
     if (!val.startsWith("/")) {
       val = "/".concat(val)
     }
@@ -74,13 +76,13 @@ function getChildren(dirPath: string, absDirPath: string): SidebarConfigArray {
   const files = fs.readdirSync(absDirPath, { withFileTypes: true })
   // 过滤文件
   const filtered = files.filter(
-    (file) =>
+    file =>
       !DEFAULT_IGNORED_FILES.find(
-        (val) => val.toLowerCase() === file.name.toLowerCase()
+        val => val.toLowerCase() === file.name.toLowerCase()
       )
   )
 
-  return filtered.map((file) => {
+  return filtered.map(file => {
     const relativeFilePath = path.join(dirPath, file.name)
     const absoluteFilePath = path.join(absDirPath, file.name)
     if (file.isDirectory()) {
@@ -126,7 +128,7 @@ function getMDFileTitle(file: string): string {
   let md_file = fs.readFileSync(file)
   let m = markdownIt().parse(md_file.toString(), {})
 
-  let h1_index = m.findIndex((val) => val.tag === "h1")
+  let h1_index = m.findIndex(val => val.tag === "h1")
   if (h1_index !== -1) {
     let h1 = m.at(h1_index + 1)
     return h1!.content
@@ -142,11 +144,11 @@ function getMDFileTitle(file: string): string {
  */
 function getHomePageFileName(absDirPath: string): string {
   let fileName = ""
-  fs.readdirSync(absDirPath, { withFileTypes: true }).forEach((file) => {
+  fs.readdirSync(absDirPath, { withFileTypes: true }).forEach(file => {
     const upperFileName = file.name.toUpperCase()
     if (
       -1 !==
-      HOME_PAGE_NAME.findIndex((val) => val.toUpperCase() === upperFileName)
+      HOME_PAGE_NAME.findIndex(val => val.toUpperCase() === upperFileName)
     ) {
       fileName = file.name
     }
@@ -154,47 +156,10 @@ function getHomePageFileName(absDirPath: string): string {
   return fileName
 }
 
-interface HeaderTokens {
-  header1: HeaderToken[]
-  header2: HeaderToken[]
-}
-
-interface HeaderToken {
-  content: string
-  map: [number, number]
-}
-
-const getMDFileTitles = (file: string): HeaderTokens => {
-  let md_file = fs.readFileSync(file)
-  let m = markdownIt().parse(md_file.toString(), {})
-
-  let header1 = getHeaders(m, "h1")
-  let header2 = getHeaders(m, "h2")
-
-  return {
-    header1,
-    header2,
-  }
-}
-
-function getHeaders(tokens: Token[], tag: string) {
-  let headers: HeaderToken[] = []
-  for (let index = 0; index < tokens.length; index++) {
-    if (tag === tokens[index].tag && tokens[index].type === "heading_open") {
-      const token = {
-        content: tokens[index + 1]?.content ?? "",
-        map: tokens[index + 1]?.map ?? [0, 0],
-      }
-      if (token) headers.push(token)
-    }
-  }
-  return headers
-}
-
 const sidebarConfig = (folders: string[]) => {
   let sidebarConfig: SidebarConfigObject = initSidebarConfigObject(folders)
   Object.keys(sidebarConfig).map(
-    (item) => (sidebarConfig[item] = getConfigArray(item))
+    item => (sidebarConfig[item] = getConfigArray(item))
   )
   return sidebarConfig
 }
