+++ tsconfig.base.json
@@ -1,0 +1,16 @@
+{
+  "extends": "tsconfig-vuepress/base.json",
+  "compilerOptions": {
+    "composite": true,
+    "lib": [
+      "DOM",
+      "ES2020"
+    ],
+    "moduleResolution": "Node",
+    "allowSyntheticDefaultImports": true,
+    "noEmitOnError": true,
+    "noImplicitAny": false,
+    "skipLibCheck": true,
+    "target": "ES2020"
+  }
+}
