+++ tsconfig.json
@@ -1,0 +1,35 @@
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "baseUrl": "./",
+    "jsx": "preserve",
+    "module": "ES2020",
+    // "paths": {
+    //   "@internal/searchIndex": [
+    //     "./packages/@vuepress/plugin-search/src/client/searchIndex"
+    //   ],
+    //   "@internal/themeData": [
+    //     "./packages/@vuepress/plugin-theme-data/src/client/themeData"
+    //   ],
+    //   "@internal/*": ["./packages/@vuepress/client/src/types/internal/*"],
+    //   "@vuepress/*": ["./packages/@vuepress/*/src"],
+    //   "vuepress": ["./packages/vuepress/src"],
+    //   "vuepress-vite": ["./packages/vuepress-vite/src"],
+    //   "vuepress-webpack": ["./packages/vuepress-webpack/src"]
+    // },
+    "types": [
+      "vite/client",
+    ]
+  },
+  "include": [
+    "**/.vuepress/**/*",
+    "packages/**/*",
+    "scripts/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    ".temp",
+    "lib",
+    "dist"
+  ]
+}
